Add support for building + linking blazesym
Add support for building and linking the blazesym library when
USE_BLAZESYM is defined. Blazesym [0] is a library that handles address
symbolization and related tasks. It can work with various formats
including ELF, DWARF, Gsym, and Breakpad and supports mapping from
address to symbol (+ meta data) as well as from symbol to address (+
meta data).

Blazesym is a Rust library and at this point we require building it from
source, meaning that a Rust toolchain will be required. In the future we
anticipate static and dynamic libraries to be shipped with more and more
distributions, which will allow us to lift this requirement and rely on
system-dependencies. To bridge the gap between Cargo (the native Rust
build manager) and CMake, we rely on Corrosion.

This change puts the logic in place for building blazesym and making it
ready for use by the program. By default (i.e., when USE_BLAZESYM is not
defined), nothing will changed, i.e., the feature is completely opt-in
right now.

[0] https://github.com/libbpf/blazesym

Signed-off-by: Daniel Müller <deso@posteo.net>