Add Config member to Ksyms type
The first piece of infrastructure harnessing blazesym for symbolization
matters will be the kernel symbolization logic in Ksyms. As motivated
earlier, to make it possible to fall back to using the BCC backend at
runtime, which is configurable via the program Config, make sure to keep
a Config instance around to consult later. Also put everything Ksyms
into the bpftrace namespace, which was missed earlier.